chore: Make dark card container darker, light more purple/theme
