+++ GPSTest/src/main/java/com/android/gpstest/ui/dashboard/ErrorCheckList.kt
@@ -72,7 +72,8 @@ fun ErrorCheckList(
 ) {
     Card(
         modifier = Modifier
-            .fillMaxWidth()
+            .fillMaxWidth(),
+        containerColor = MaterialTheme.colorScheme.primaryContainer
     ) {
         Column(
             modifier = Modifier.padding(leftColumnMargin)
