+++ GPSTest/src/main/java/com/android/gpstest/ui/dashboard/FeaturesAccuracyList.kt
@@ -67,7 +67,8 @@ fun FeaturesAccuracyList(
 ) {
     Card(
         modifier = Modifier
-            .fillMaxWidth()
+            .fillMaxWidth(),
+        containerColor = MaterialTheme.colorScheme.primaryContainer
     ) {
         Column(
             modifier = Modifier.padding(leftColumnMargin)
